Fixed dice coef. return value on emtpy sets (#111)
* Fixed dice coef. return value on emtpy sets

While not clearly defined, the convention seems to be that an empty result and reference set should cause the dice coefficient to return `1` (i.e. a perfect result) rather than `0`. Adapted accordingly.

* Fixed Jacard ZeroDivision behaviour

Would have thrown a ZeroDivisionError, but correct behavior is rather returning a value of `1` (i.e. perfect fit).